Simplifying CodeLibrary backend
